requirements.txt
aiodns==4.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
autoflake==2.3.1
backoff==2.2.1
bcrypt==5.0.0
black==26.1.0
boto3==1.42.49
botocore==1.42.49
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coverage==7.13.4
cryptography==45.0.7
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
execnet==2.1.2
fastapi==0.129.0
fastapi-cli==0.0.23
fastapi-cloud-cli==0.12.0
fastar==0.8.0
filelock==3.24.2
frozenlist==1.8.0
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
inquirerpy==0.3.4
invoke==2.2.1
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
librt==0.8.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
mypy==1.19.1
mypy_extensions==1.1.0
numpy==2.4.2
openai==2.21.0
orjson==3.11.7
packaging==26.0
paramiko==4.0.0
pathspec==1.0.4
pfzy==0.3.4
platformdirs==4.9.2
pluggy==1.6.0
portalocker==3.2.0
prettytable==3.17.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
py-cpuinfo==9.0.0
pycares==5.0.1
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.0
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyNaCl==1.6.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-mock==3.15.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytokens==0.4.1
PyYAML==6.0.3
qdrant-client==1.16.2
requests==2.32.5
responses==0.25.8
rich==14.3.2
rich-toolkit==0.19.4
rignore==0.7.6
runpod==1.8.1
s3transfer==0.16.0
sentry-sdk==2.53.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
tomli==2.4.0
tomlkit==0.14.0
tqdm==4.67.3
tqdm-loggable==0.3
typer==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
yarl==1.22.0
